Test client sending flatbuffer Handshake
This adds creating the Handshake message in the client, and adjusts dcas to
be able to properly recieve it. I had to remove some of the old text parsing
code from the dcas loop because it quit on recieving the now-binary data
from the test client. dcas now processes the incoming flatbuffer as binary
and is able to recognize the incoming session handshake message.